requirements-compatible.txt
# Compatible package versions for Python 3.12
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Session==0.8.0
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0

# Data science packages - newer versions compatible with Python 3.12
pandas>=2.1.0
numpy>=1.26.0
joblib>=1.3.0
scikit-learn>=1.3.0

# Image processing - newer versions
Pillow>=10.0.0
opencv-python>=4.8.0

# Socket.IO
python-socketio>=5.8.0

# PyTorch - CPU version for compatibility
torch>=2.1.0
torchvision>=0.16.0

# ML libraries - newer versions
xgboost>=2.0.0
lightgbm>=4.1.0
catboost>=1.2.0

# Google AI
google-generativeai>=0.8.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
pydicom>=2.4.0

# WSGI server
waitress>=3.0.0

# Optional AI packages (install if available)
segmentation-models-pytorch
ultralytics

# OCR (optional)
pytesseract